Migrate code style to Black
This project does not follow any particular code style conventions, which is a
distraction to some (me, in particular), but has the general effect of making
the code harder to read and edit. As an example, some modules use tabs for
indentation and others use spaces. Good code style depends entirely on
programmers' individual choices.

[Black] is an opinionated code formatter with intentionally limited
configuration options, so that developers spend less time thinking about code
style. The intention is for Black to make the vast majority of code style
decisions and make a consistent style more likely throughout the project.

[Black]: https://black.readthedocs.io/en/stable/index.html