Return dict representing RapidPro flow from create_flows function
The decision to write files to the file system should be left to the caller. We
always return a dict in case further processing of the flows is required after
conversion from source spreadsheets.

The `output_files` argument is deprecated and will be removed in future. Anyone
calling `create_flows` should pass in `None` for `output_files` and write a file
themselves, if they wish.