Refactor CSV sheet reader to manage a directory of CSV files
It simplifies things a lot to define a CSV workbook to be a directory containing
CSV files.

```
workbook
├── content_index.csv
├── sheet1.csv
└── sheet2.csv
```

This should be considered the same as an XLSX file or Google Sheets document
with three sheets/tabs. A CSV reader can be identified by its base directory, so
two CSV readers with the same base directory are identical.