Create surveys feature (#147)
RapidPro contact fields have a name and a key.
In the UI, users enter a name, and it is converted to a key (which is all
lowercase and has spaces converted to underscores).
The key is used everywhere when we want to reference a variable, e.g. via
`@fields.key`. The name is used for display, but otherwise is pretty useless.
Yet, RapidPro enforces that the name may not contain underscores, and thus it
must differ from the key if we want to have underscores in the key.
This is cumbersome.

Therefore, in flow definitions, we now allow specifying a `save_name` for
`save_value` rows which may contain underscores. Thus the exact same string can
be used both in this column and when referencing the field via `@fields.key`.
A name is autogenerated by replacing the underscores with spaces (but for a
flow author, the name is irrelevant anyway).

---------

Co-authored-by: Ian Stride <ianstride@gmail.com>