Remove shutdown handler (#157)
Removes the ShutdownHandler logging handler in favour of raising and (optionally) catching exceptions. The behaviour of the toolkit should be exactly the same as before - making this change backwards-compatible.

The shutdown handler used the Python logging framework to control program execution, in particular, halting the program if a critical-level log message is emitted. This is undesirable for the following reasons:

Surprise factor - this is not a widespread practice, and it is not obvious or expected that logging a critical message should halt the program.

Testing critical conditions is more difficult because there is no obvious way to test for the presence of a log message, but it is easy to test whether an exception has been raised.

Clients of the toolkit (e.g. pipeline, CLI) should be free to configure their own logging preferences without (inadvertently) affecting the execution of the toolkit