Support for global templating context (#197)
* Use enum in content index parser

* Support for global context

* rename globals_dict to globals

Make compatible with universal parser
Wrap MockSheetReader in source object, as ContentIndexParser only takes sources - point in favor of having strong type hinting
Removing _get_sheet_or_die, as that is now covered in the data source class. Possibly that method should be a public method for cases like this.

* Get global properties via data_source.get

---------

Co-authored-by: GabeBolton <jb.gabe@gmail.com>
Co-authored-by: Ian Stride <ianstride@gmail.com>